chore(fl): move ui.h aggregator into fl/ui/ subdirectory
Relocates `src/fl/ui.h` to `src/fl/ui/ui.h` so the aggregator lives
beside the per-element headers it re-exports. Updates all 44 consumers
(examples, src, tests, docs) to use the new include path. Test file
`tests/fl/ui.cpp` is moved to `tests/fl/ui/ui.cpp` to satisfy the
TestPathStructureChecker mirror requirement.
