+++ examples/AnimartrixRing/AnimartrixRing.ino
@@ -13,7 +13,7 @@
 
 #include "fl/math/math.h"
 #include "fl/math/screenmap.h"
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/math/math.h"
 #include "fl/fx/2d/animartrix.hpp"
 #include "fl/audio/audio_processor.h"
