+++ examples/Audio/simple/simple.h
@@ -32,7 +32,7 @@ all the UI elements you see below.
 #include "fl/math/math.h"
 #include "fl/gfx/raster.h"
 #include "fl/math/time_alpha.h"
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/gfx/xypath.h"
 #include "fl/stl/compiler_control.h"
 #include "fl/fx/time.h"
