+++ examples/AudioUrl/AudioUrl.ino
@@ -12,7 +12,7 @@
 
 #include <FastLED.h>
 #include "fl/asset/asset.h"
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/stl/url.h"
 
 #define NUM_LEDS 64
