+++ examples/Blur2d/Blur2d.ino
@@ -15,7 +15,7 @@
 #include <Arduino.h>
 #include <FastLED.h>
 
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/math/xymap.h"
