+++ examples/Chromancer/Chromancer.ino
@@ -34,7 +34,7 @@
 #include "fl/math/screenmap.h"
 #include "fl/math/math.h"
 #include "fl/stl/json.h"
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/stl/map.h"
 
 #include "fl/stl/string.h"
