+++ examples/FireCylinder/FireCylinder.h
@@ -34,7 +34,7 @@ with colors transitioning from black to red/yellow/white (or other palettes).
 // 5. The time dimension adds continuous variation to make the fire look dynamic
 
 #include "FastLED.h"     // Main FastLED library for controlling LEDs
-#include "fl/ui.h"       // UI components for the FastLED web compiler (sliders, buttons, etc.)
+#include "fl/ui/ui.h"       // UI components for the FastLED web compiler (sliders, buttons, etc.)
 #include "fl/math/xymap.h"    // Mapping between 1D LED array and 2D coordinates
 #include "fl/fx/time.h"     // Time manipulation utilities for animations
