+++ examples/FireMatrix/FireMatrix.h
@@ -39,7 +39,7 @@ void loop() {}
 #else
 
 #include "FastLED.h"     // Main FastLED library for controlling LEDs
-#include "fl/ui.h"       // UI components for the FastLED web compiler (sliders, etc.)
+#include "fl/ui/ui.h"       // UI components for the FastLED web compiler (sliders, etc.)
 #include "fl/math/xymap.h"    // Mapping between 1D LED array and 2D coordinates
 #include "fl/fx/time.h"     // Time manipulation utilities
