+++ examples/XYPath/complex.h
@@ -20,7 +20,7 @@ all the UI elements you see below.
 #include "fl/math/math.h"
 #include "fl/gfx/raster.h"
 #include "fl/math/time_alpha.h"
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/gfx/xypath.h"
 #include "fl/fx/time.h"
