+++ examples/wasm/wasm_impl.h
@@ -18,7 +18,7 @@
 #include "fl/fx/fx_engine.h"
 
 #include "fl/fx/2d/animartrix.hpp"
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
