+++ src/fl/sensors/digital_pin.cpp.hpp
@@ -1,7 +1,7 @@
 
 #include "fl/stl/stdint.h"
 
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/stl/shared_ptr.h"  // For make_shared
 
 #include "fl/sensors/digital_pin.h"
