+++ tests/fl/audio_url.cpp
@@ -2,7 +2,7 @@
 // Tests that UIAudio with a URL properly serializes the URL field
 // in the JSON UI output on the stub platform.
 
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "fl/stl/json.h"
 #include "fl/system/log.h"
 #include "fl/stl/url.h"
