+++ tests/fl/ui/ui.cpp
@@ -1,7 +1,7 @@
 // g++ --std=c++11 test.cpp
 
 
-#include "fl/ui.h"
+#include "fl/ui/ui.h"
 #include "platforms/shared/ui/json/ui.h"
 #include "platforms/shared/ui/json/ui_internal.h"
 #include "platforms/shared/ui/json/number_field.h"
